refactor: 타입 파일 이름변경에 따른 import 수정
