+++ src/app/(header)/(side-bar)/researches/projects/[id]/edit/page.tsx
@@ -4,7 +4,7 @@ import { useRouter } from 'next/navigation';
 import { ProjectForm } from '@/components/researches/projects/project-form';
 import { Button } from '@/components/ui/button';
 import { ArrowLeft } from 'lucide-react';
-import { Project, ResearchCategory, ResearchStatus } from '@/types/project';
+import { Project, ProjectCategory, ResearchStatus } from '@/types/researches';
 
 // 임시 데이터
 const project = {
@@ -15,7 +15,7 @@ const project = {
     '본 연구는 최신 생물정보학 기술을 활용하여 유전체 데이터를 분석하고 질병 관련 유전자 변이를 식별하는 것을 목표로 합니다.\n\n한국인 유전체 데이터의 특성을 고려한 분석 모델 개발에 중점을 두고 있으며, 대규모 유전체 데이터 기반 머신러닝 모델을 개발하고 임상 검증을 설계합니다.',
   startDate: '2023-01-01',
   endDate: '2023-12-31',
-  category: 'Bioinformatics' as ResearchCategory,
+  category: 'Bioinformatics' as ProjectCategory,
   status: '진행 중' as ResearchStatus,
   createdAt: '2023-01-01T09:00:00Z',
   updatedAt: '2023-05-01T10:00:00Z',
@@ -39,7 +39,7 @@ export default function EditProject({ params }: { params: { id: string } }) {
   };
 
   return (
-    <div className="container">
+    <div className="flex flex-col px-30">
       <div className="mb-6 flex items-center">
         <Button
           variant="ghost"
